"Fixed scrollview wouldn't scroll all of the way to the top. This is apparently some oddity or misunderstanding with the gravity attribute.
Here's what I had originally. I have my FlexboxLayout within my ScrollView. In order to get it to center vertically, I set the gravity=center attribute on the FlexboxLayout. That did center it, but then I got this problem where I couldn't scroll all the way to the top.

I googled it and another guy had a similar problem on stackoverflow. The
answer was to remove gravity=center, and on the ScrollView, add
fillViewport=true. So I did that, but I think that was making my dice
stretch all the way vertically, to fill the ScrollView. So eventually I
tried setting the alignItems=center attribute in the FlexboxLayout, and
I think that fixed it. And then I set the alignContent=center, too, so
that the rows would be bunched up in the center. I think, by default,
they were sort of evenly spacing out from the center.